safari-534.53-branch/LayoutTests/fast/js/string-capitalization-expected.txt

This test checks that toLowerCase and toUpperCase handle certain non-trivial cases correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS String("Ağ€").toLowerCase() is "ağ¨"
PASS String("ağ¨").toUpperCase() is "Ağ€"
PASS String("ÎšÎŸÎ£ÎœÎŸÎ£ ÎšÎŸÎ£ÎœÎŸÎ£").toLowerCase() is "ÎºÎ¿ÏƒÎ¼Î¿Ï‚ ÎºÎ¿ÏƒÎ¼Î¿Ï‚"
PASS String("ÃŸ").toUpperCase() is "SS"
PASS String("Å‰").toUpperCase() is "Ê¼N"
PASS String("Ç°").toUpperCase() is "JÌŒ"
PASS String("ï¬ƒ").toUpperCase() is "FFI"
PASS String("FFI").toLowerCase() is "ffi"
PASS String("Ä²").toLowerCase() is "Ä³"
PASS successfullyParsed is true

TEST COMPLETE
